+++ src/manga_archiver/utils/__init__.py
@@ -1,9 +1,11 @@
+from .download_limiter import DownloadLimiter
 from .downloader import DownloadClient
 from .logger import setup_logging
 from .multi_format_exporter import MultiFormatExporter
 
 __all__ = [
     "DownloadClient",
+    "DownloadLimiter",
     "MultiFormatExporter",
     "setup_logging",
 ]